+++ src/uk/ac/cam/cl/juliet/test/messagedecoding/XDPProcessingTest.java
@@ -275,8 +275,8 @@ class TestDatabase implements DatabaseConnection {
 	}
 
 	@Override
-	public ResultSet getTradesInTimeRangeForSymbol(long symbolID, int start,
-			int end) throws SQLException {
+	public ResultSet getTradesInTimeRangeForSymbol(long symbolID, long start,
+			long end) throws SQLException {
 		return null;
 	}
